+++ src/evolduo_app/mailer.clj
@@ -33,6 +33,8 @@
 (defn evolution-url [app-url evolution-id]
   (str app-url "/evolution/" evolution-id))
 
+(defn- password-reset-url [app-url token]
+  (str app-url "/user/set-password-form?token=" token))
 
 (defn email-template [{:keys [app-url company-address] :as settings} {:keys [unsubscribe_token] :as user} {:keys [title content]}]
   [:html
@@ -190,13 +192,18 @@ table.body .article {
     [(p (str "User " invited-by-email " invited you to collaborate on the following track"))
      (a {:href (evolution-url app-url evolution-id) :text "View"})]))
 
-(defn- get-email-data [db {:keys [app-url] :as settings} {:keys [verification_token subscription] :as user} mail]
+(defn- get-email-data [db {:keys [app-url] :as settings} {:keys [verification_token subscription password_reset_token] :as user} mail]
   (condp = (:type mail)
     "signup" {:should-receive? true :title "Welcome to Evolduo"
               :content         (email-template settings user {:title "Welcome to Evolduo"
                                                               :content [(p "Welcome to Evolduo")
-                                                                      (p "Please click the following link to verify your account")
-                                                                      (a {:href (verification-url app-url verification_token) :text "Verify"})]})}
+                                                                        (p "Please click the following link to verify your account")
+                                                                        (a {:href (verification-url app-url verification_token) :text "Verify"})]})}
+    "password-reset" {:should-receive? true :title "Reset your Evolduo password"
+                      :content         (email-template settings user {:title "Reset your Evolduo password"
+                                                                      :content [(p "Hi,")
+                                                                                (p "If you requested a password reset please click the link below to reset your password.")
+                                                                                (a {:href (password-reset-url app-url password_reset_token) :text "Reset Password"})]})}
     "invitation" {:should-receive? (:notifications subscription)
                   :title "Invitation to collaborate" ;; TODO should receive, duplication
                   :content         (email-template settings user {:title "Invitation to collaborate"